Move get / poll mutable state to api package
